Configuration script for alias and hook
This script copies the pre-commit hook in .env-config to all
repositories included in the meta repo. It also adds an alias to a
shell's default configuration file for the google style formatter JAR.
It's important to note that not all shells can be supported. Only the
primary shells expected to be on the system are targeted.